### 2023-02-16 #574 @joelnitta
dwctaxon: Tools for Working with Darwin Core Taxon Data
Submitting Author Name: Joel H. Nitta
Submitting Author Github Handle: @joelnitta
Repository:  https://github.com/joelnitta/dwctaxon
Version submitted: 1.0.0.9000
Submission type: Standard

Version accepted: TBD
---

```
Package: dwctaxon
Title: Tools for Working with Darwin Core Taxon Data
Version: 1.0.0.9000
    c(
      person(given = "Joel H.",
           family = "Nitta",
           email = "joelnitta@gmail.com",
      person(given = "Wataru",
           family = "Iwasaki",
           role = c("ctb"),
    )
Description: Provides functions to create, manipulate, and validate taxonomic
    data in compliance with Darwin Core standards
    (Darwin Core "Taxon" class https://dwc.tdwg.org/terms/#taxon).
Roxygen: list(
    markdown = TRUE,
    roclets = c("collate", "namespace", "rd", "roxyglobals::global_roclet"))
    assertr,
    assertthat,
    digest,
    glue,
    settings,
    utils
    roxyglobals (>= 0.2.1),
    mockery,
    usethis,
    tidyverse,
    patrick,
    stringi
    anthonynorth/roxyglobals
    https://github.com/joelnitta/dwctaxon
```

dwctaxon facilitates manipulating and validating taxonomic data (of biological species) in R, in compliance with the widely used Darwin Core data standard.

Biologists, anybody working with taxonomic data in Darwin Core format.

Not currently, to the best of my knowledge. The closest thing out there is the [GBIF data validator](https://www.gbif.org/tools/data-validator) (not an R package). The archived finch package [had a function to call the GBIF data validator](https://github.com/ropensci-archive/finch/issues/29).

No pre-submission inquiry

pkgcheck passes when I run it locally. I have observed an error in the CI, but I believe [this is a problem with pkgcheck, and unrelated to my package](https://github.com/ropensci-review-tools/pkgcheck-action/issues/34).
