### 2023-02-16 joelnitta/dwctaxon#43 @joelnitta
Check for non-used imports
detected here: https://github.com/ropensci/software-review/issues/574#issuecomment-1432893746, in particular

assertthat, digest, rlang
