### 2023-03-01 #574 @noamross
Thanks @collinschwantes! Please be sure to look at the diagnostic report above.  I note a few instances of packages imported for very few functions (assertr), high-cyclocomplexity functions that may benefit from breaking up, and some other `goodpractice` notes that should be examined as part of your review.
