+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   editing: editing.html
   validation: validation.html
   what-is-dwc: what-is-dwc.html
-last_built: 2023-03-21T00:05Z
+last_built: 2023-03-21T21:52Z
 urls:
   reference: https://joelnitta.github.io/dwctaxon/reference
   article: https://joelnitta.github.io/dwctaxon/articles
