### 2023-04-17 #574 @sformel-usgs
- **Briefly describe any working relationship you have (had) with the package authors:** None

---

I found this package to be thoughtfully written and documented. Code appears to have been carefully constructed and functions well.  However, I do have a few high level concerns:

1. TL;DR More context and example should be added to the statement of need.

   - **The statement of need** explains what you can do with the package, but doesn’t do a good job justifying a need or specifying the use cases and gaps the package fills. My first reaction, and that of a colleague I shared the readme with, was that we couldn’t clearly understand why this package was needed from the readme alone. However, I found this [talk given by the author](https://www.youtube.com/watch?v=NUM3tFCF5XY) to do an excellent job justifying the package creation and explaining the circumstances in which you would want to use it. It may be enough to link to the video as a ‘more detail’ resource. If not, more is needed to explain why/when the need for the package exists. To give a little more context, I am the type of user that is described at the end of the video (mostly referencing authorities like WoRMS, rarely building my own databases). That being said, after working through the vignettes, I was able to imagine how the package would be useful.

2.  I’m not sure why these functions shouldn’t be included as part of the `taxastand` package (or vice versa)? Again, the video referenced above give some sense of why these tasks would be separated, but I think a little more description of how they relate could be useful. I’m not totally convinced that the modification/validation of taxonomic data (`dwctaxon`) should be separated from the standardization of species names from different sources (`taxastand`) as these tasks seem to have a lot of overlap.

Generally, the instructions and documentation were clear, thoughtful and useful. It made good use of the principle of multiple points of entry without sprawling unnecessarily. I especially appreciate that `dct_options` is highlighted in the [validation vignette](https://joelnitta.github.io/dwctaxon/articles/validation.html#changing-the-defaults), as I think the variety of outputs that can be returned by the functions is a strength of this package.  I would also like to compliment the author on the “What is DWC?” vignette. It is a thorough and clear explanation of DwC with regard to taxonomy and will be helpful to novices. I do, however, have some suggestions below.

1.  Throughout the documentation, but most noticeably in the vignette [‘What is DWC?’ vignette](https://github.com/joelnitta/dwctaxon/blob/main/vignettes/what-is-dwc.Rmd), Darwin Core is abbreviated ‘DWC’. Generally speaking, I see/use ‘DwC’, and that is what is described on the [Wikipedia page](https://en.wikipedia.org/wiki/Darwin_Core) since 2011. Not critical, but I would be curious to hear whether the author has a strong opinion on the appropriate abbreviation. If not, I would prefer DwC.

2.  In the [‘What is DWC?’ vignette](https://github.com/joelnitta/dwctaxon/blob/main/vignettes/what-is-dwc.Rmd), line 64 describes `genus`, `family` and `order` as “higher taxonomic levels”. I understand these to be “lower levels”, which seems consistent with the comment in the DwC taxonomic term [higherClassification](https://dwc.tdwg.org/terms/#dwc:higherClassification).

3.  Line 64 of the [editing vignette](https://github.com/joelnitta/dwctaxon/blob/main/vignettes/editing.Rmd) references ‘hash’ generically. I think it would be worth specifying that it is using MD5.

4. `utils.R` credits the `paste3` function to a stackoverflow conversation. I’m not sure of the best way to give credit in this situation, but it seems like it would also be good to include the user \[“IRTFM”\] and userID \[“1855677”\], since it is a function that is clearly written by one person.

#### Logical programming API

1.  Yes, however see suggestions below.
2.  In the function, `dct_add_row`, I’m curious about the decision to only use the first 8 characters from the hash as the taxonID. As you note, this may result in duplicates, why not preserve the entire hash as taxonID?
3.  I disagree with the use/creation of aliases as described in lines 72-89 of the [editing vignette](https://github.com/joelnitta/dwctaxon/blob/main/vignettes/editing.Rmd). This promotes confusion about the name of the DwC standard term. Especially in this case where the term is a function parameter and users can use tab to autocomplete, rather than typing out the names. I think this aspect of the package should be revised, and the vignette can point out tab autocomplete as a way to avoid typing out the rather long names of DwC.

#### automated tool review

Checking the initial package report generated by our @ropensci-review-bot.

1.  As noted in the `goodpractice` checks, `dct_modify_row` and `dct_add_row` have high-cyclocomplexity. I don’t have any great suggestions for ways to break it up. But, if my suggestion, that aliases for DwC terms should be avoided, is accepted, then some of the checks for misuse of aliases could be removed.
2.  The `library` and `require` calls described by `lintr` in `goodpractices` are acceptable. They all appear to be associated with vignettes, tests, or documentation of raw data creation.
3.  It appears that the `assertr` package is imported for a single function in `utilties.R`. I don’t know of a simple way to replace this function and reduce the dependency.
4.  The package dependency analysis did not identify calls for the import packages, `assert_that`, `digest`, and `rlang`. I have checked that these are necessary for the package to function.

Checking the package’s logs on its continuous integration services (GitHub Actions, Codecov, etc.)

1.  One issue addressed by the author as a possible problem with pkgcheck when run remotely.

Running devtools::check() and devtools::test() on the package to find any errors that may be missed on the author’s system.

1.  Passed without error or unusual skips, and warnings.
