### 2023-04-25 #574 @collinschwantes
-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

-   [] **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README ~

Estimated hours spent reviewing: ~8

-   [x] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

Overall easy to use and well documented. I really enjoyed testing this package.
That being said, the statement of need could more clearly define who the target audience is.
Is it  biologist who are not necessarily taxonomists, taxonomists, collections managers?

I tested the package using data from a project I worked on back in 2014. Once I had the taxa data
in the proper form, all functions worked as expected. I did notice that certain functions create missing columns or provide errors for non-dct terms. Its not always obvious which functions will generate new columns, and which wont. See `dct_fill_col` vs `dct_add_row`

- I would prefer that functions use DCW terms instead of abbreviations.
I didn't find the terms saved me that much time from a typing perspective
and I had to remember the abbreviation plus the actual term. Not a lot of mental effort
but non-zero.
- Since there is an "add_row" function, it may be worth including a drop_row?
- `dct_fill_col` does not throw an error if the `fill_from` argument is a non-existent column that is a dct_term.
- One error for failing test when running `devtools::check`. `dct_add_row` throws an error.
- dct_modify_row_single has a lot of if statements that are nicely labelled in the function. Consider whether or not each of those sections of the function could be split into its own, smaller, easier to evaluate function.
- You could also take advantage of `rlang::empty` to test if something is NA or
NULL.
- The dependency section in the packaging guidelines recommends using minimum version numbers
for the package dependencies
- Would be great (though potentially out of scope for the package) to have a
vignette that shows how to retrieve taxon data and/or what
you can do with taxon data once you have it properly formatted in Darwin core.
