### 2023-04-25 joelnitta/dwctaxon#44 @joelnitta
Allow longer hashes for taxonID
As suggested by @sformel-usgs in https://github.com/ropensci/software-review/issues/574#issuecomment-1512072208

> In the function, dct_add_row, I’m curious about the decision to only use the first 8 characters from the hash as the taxonID. As you note, this may result in duplicates, why not preserve the entire hash as taxonID?
