+++ DESCRIPTION
@@ -22,7 +22,7 @@ LazyData: true
 Roxygen: list(
     markdown = TRUE,
     roclets = c("collate", "namespace", "rd", "roxyglobals::global_roclet"))
-RoxygenNote: 7.2.1
+RoxygenNote: 7.2.3
 Imports: 
     assertr,
     assertthat,
