### 2023-04-25 ef7c393
Allow longer hashes for taxonID fixes #44
