+++ man/dct_add_row.Rd
@@ -14,6 +14,7 @@ dct_add_row(
   new_dat = NULL,
   fill_taxon_id = dct_options()$fill_taxon_id,
   fill_usage_id = dct_options()$fill_usage_id,
+  taxon_id_length = dct_options()$taxon_id_length,
   stamp_modified = dct_options()$stamp_modified,
   strict = dct_options()$strict,
   ...
@@ -43,9 +44,11 @@ the above arguments, except for \code{tax_dat}. Other DWC terms can also be
 included as additional columns. All rows in \code{new_dat} will be appended to the
 input data (\code{tax_dat}).}
 
-\item{fill_taxon_id}{Logical vector of length 1; if \code{taxon_id} is not provided, should values in the taxonID column be filled in by looking them up from the scientificName? Default \code{TRUE}.}
+\item{fill_taxon_id}{Logical vector of length 1; if \code{taxon_id} is not provided, should values in the taxonID column be filled in by generating them automatically from the scientificName? Default \code{TRUE}.}
 
 \item{fill_usage_id}{Logical vector of length 1; if \code{usage_id} is not provided, should values in the acceptedNameUsageID column be filled in by matching acceptedNameUsage to scientificName? Default \code{TRUE}.}
+
+\item{taxon_id_length}{Numeric vector of length 1; how many characters should be included in automatically generated values of taxonID? Must be between 1 and 32, inclusive. Default \code{32}.}
 
 \item{stamp_modified}{Logical vector of length 1; should the \code{modified} column of any newly created or modified row include a timestamp with the date and time of its creation/modification? Default \code{TRUE}.}
 
@@ -70,6 +73,10 @@ are provided as convenience to avoid typing the longer "taxonomicID",
 
 \code{fill_taxon_id} and \code{fill_usage_id} only act on the newly added data (they
 do not fill columns in \code{tax_dat}).
+
+If "taxonID" is not provided for the new row and \code{fill_taxon_id} is \code{TRUE},
+a value for taxonID will be automatically generated from the md5 hash digest
+of the scientific name.
 
 To modify settings used for validation if \code{strict} is \code{TRUE},
 use \code{dct_options()}.
