### 2023-04-25 joelnitta/dwctaxon#46 @joelnitta
Statement of need is not sufficiently clear
As mentioned in [ROpenSci review](https://github.com/ropensci/software-review/issues/574)

@collinschwantes:

> the statement of need could more clearly define who the target audience is. Is it biologist who are not necessarily taxonomists, taxonomists, collections managers?

@sformel-usgs：

> The statement of need explains what you can do with the package, but doesn’t do a good job justifying a need or specifying the use cases and gaps the package fills. My first reaction, and that of a colleague I shared the readme with, was that we couldn’t clearly understand why this package was needed from the readme alone. However, I found this [talk given by the author](https://www.youtube.com/watch?v=NUM3tFCF5XY) to do an excellent job justifying the package creation and explaining the circumstances in which you would want to use it. It may be enough to link to the video as a ‘more detail’ resource. If not, more is needed to explain why/when the need for the package exists. To give a little more context, I am the type of user that is described at the end of the video (mostly referencing authorities like WoRMS, rarely building my own databases). That being said, after working through the vignettes, I was able to imagine how the package would be useful.
