### 2023-04-25 joelnitta/dwctaxon#47 @joelnitta
Some parts of the vignettes are unclear ('DwC', 'higher taxonomic levels', 'hash')
As mentioned in [ROpenSci review](https://github.com/ropensci/software-review/issues/574) by @sformel-usgs:

> Throughout the documentation, but most noticeably in the vignette [‘What is DWC?’ vignette](https://github.com/joelnitta/dwctaxon/blob/main/vignettes/what-is-dwc.Rmd), Darwin Core is abbreviated ‘DWC’. Generally speaking, I see/use ‘DwC’, and that is what is described on the [Wikipedia page](https://en.wikipedia.org/wiki/Darwin_Core) since 2011. Not critical, but I would be curious to hear whether the author has a strong opinion on the appropriate abbreviation. If not, I would prefer DwC.

> In the [‘What is DWC?’ vignette](https://github.com/joelnitta/dwctaxon/blob/main/vignettes/what-is-dwc.Rmd), line 64 describes genus, family and order as “higher taxonomic levels”. I understand these to be “lower levels”, which seems consistent with the comment in the DwC taxonomic term [higherClassification](https://dwc.tdwg.org/terms/#dwc:higherClassification).

> Line 64 of the [editing vignette](https://github.com/joelnitta/dwctaxon/blob/main/vignettes/editing.Rmd) references ‘hash’ generically. I think it would be worth specifying that it is using MD5.
