### 2023-04-25 joelnitta/dwctaxon#48 @joelnitta
Proper attribution missing for code originating from SO
As mentioned in [ROpenSci review](https://github.com/ropensci/software-review/issues/574) by @sformel-usgs:

> utils.R credits the paste3 function to a stackoverflow conversation. I’m not sure of the best way to give credit in this situation, but it seems like it would also be good to include the user [“IRTFM”] and userID [“1855677”], since it is a function that is clearly written by one person.
