### 2023-04-25 joelnitta/dwctaxon#49 @joelnitta
Aliases should not be used for DwC terms
As mentioned in [ROpenSci review](https://github.com/ropensci/software-review/issues/574).

@sformel-usgs:

> I disagree with the use/creation of aliases as described in lines 72-89 of the [editing vignette](https://github.com/joelnitta/dwctaxon/blob/main/vignettes/editing.Rmd). This promotes confusion about the name of the DwC standard term. Especially in this case where the term is a function parameter and users can use tab to autocomplete, rather than typing out the names. I think this aspect of the package should be revised, and the vignette can point out tab autocomplete as a way to avoid typing out the rather long names of DwC.

@collinschwantes:

> I would prefer that functions use DCW terms instead of abbreviations.
I didn't find the terms saved me that much time from a typing perspective
and I had to remember the abbreviation plus the actual term. Not a lot of mental effort
but non-zero.
