### 2023-04-25 joelnitta/dwctaxon#50 @joelnitta
dct_modify_row and dct_add_row have high-cyclocomplexity
As mentioned in [ROpenSci review](https://github.com/ropensci/software-review/issues/574) by @sformel-usgs:

> As noted in the goodpractice checks, dct_modify_row and dct_add_row have high-cyclocomplexity. I don’t have any great suggestions for ways to break it up. But, if my suggestion, that aliases for DwC terms should be avoided, is accepted, then some of the checks for misuse of aliases could be removed.
