### 2023-04-25 joelnitta/dwctaxon#51 @joelnitta
dct_fill_col() mistakenly allows use of a missing column for fill_from
As mentioned in [ROpenSci review](https://github.com/ropensci/software-review/issues/574) by @collinschwantes:

> `dct_fill_col` does not throw an error if the `fill_from` argument is a non-existent column that is a `dct_term`.
