### 2023-04-25 joelnitta/dwctaxon#52 @joelnitta
DESCRIPTION lacks minimum package version numbers
As mentioned in [ROpenSci review](https://github.com/ropensci/software-review/issues/574) by @collinschwantes:

> The dependency section in the packaging guidelines recommends using minimum version numbers for the package dependencies
