### 2023-04-25 joelnitta/dwctaxon#53 @joelnitta
Should add vignette showing how to fetch data
As mentioned in [ROpenSci review](https://github.com/ropensci/software-review/issues/574) by @collinschwantes:

> Would be great (though potentially out of scope for the package) to have a vignette that shows how to retrieve taxon data and/or what you can do with taxon data once you have it properly formatted in Darwin core.
