### 2023-04-26 7d4edf4
Add checks for existence of columns in dct_fill_col
fixes #51
