### 2023-04-26 3f52a9c
Add checks for existence of columns in dct_fill_col
fixes #51
