+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   editing: editing.html
   validation: validation.html
   what-is-dwc: what-is-dwc.html
-last_built: 2023-04-25T11:03Z
+last_built: 2023-04-26T11:04Z
 urls:
   reference: https://joelnitta.github.io/dwctaxon/reference
   article: https://joelnitta.github.io/dwctaxon/articles
