### 2023-04-27 c866731
Update README
adds more detailed Statement of Need, fixes #46
