### 2023-04-27 03357f3
Update README
adds more detailed Statement of Need, fixes #46
