### 2023-04-28 3d97a19
Provide attribution for code from SO fixes #48
