+++ R/utils.R
@@ -5,27 +5,24 @@
 #' Removes NAs from pasted elements, but if ALL elements are NA, the result is
 #' NA.
 #'
-#' Shamelessly copied from
+#' Code by user IRTFM (stack overflow ID 1855677)
+#' copied from
 #' \url{https://stackoverflow.com/questions/13673894/suppress-nas-in-paste}
+#' under
+#' [CC BY-SA 3.0](https://creativecommons.org/licenses/by-sa/3.0/).
+#'
 #' @param ... Strings to paste
 #' @param sep Character used to separate pasted strings
 #' @noRd
 #' @autoglobal
 paste3 <- function(..., sep = " ") {
-  my_list <- list(...)
-  my_list <- lapply(my_list, function(x) {
-    x[is.na(x)] <- ""
-    x
-  })
-  ret <- gsub(
-    paste0("(^", sep, "|", sep, "$)"), "",
-    gsub(
-      paste0(sep, sep), sep,
-      do.call(paste, c(my_list, list(sep = sep)))
-    )
-  )
-  is.na(ret) <- ret == ""
-  ret
+  L <- list(...) # nolint
+     L <- lapply(L,function(x) {x[is.na(x)] <- ""; x}) # nolint
+     ret <-gsub(paste0("(^",sep,"|",sep,"$)"),"", # nolint
+                 gsub(paste0(sep,sep),sep, # nolint
+                      do.call(paste,c(L,list(sep=sep))))) # nolint
+     is.na(ret) <- ret=="" # nolint
+     ret
 }
 
 #' Drop the first element of a vector
