### 2023-04-28 5ac0a49
Add git2r to .pre-commit-config.yaml per https://results.pre-commit.ci/run/github/434126221/1682720588.HnwFGQnFQt2Hb40RnHL0_g
