+++ README.md
@@ -76,7 +76,7 @@ dct_filmies
 #>  8 54133788 54115100            synonym         species   Cephalomanes oblongifolium Presl          
 #>  9 54133789 54115100            synonym         species   Cephalomanes zollingeri V. D. Bosch       
 #> 10 54133790 54115100            synonym         species   Lacostea javanica (Bl.) Prantl            
-#> # … with 2,441 more rows
+#> # ℹ 2,441 more rows
 ```
 
 `dct_filmies` is a taxonomic dataset of filmy ferns included in
@@ -102,14 +102,14 @@ filmies_small |>
     taxonomicStatus = "accepted"
   )
 #> # A tibble: 6 × 6
-#>   taxonID  acceptedNameUsageID taxonomicStatus taxonRank scientificName                            modified           
-#>   <chr>    <chr>               <chr>           <chr>     <chr>                                     <chr>              
-#> 1 54115096 <NA>                accepted        species   Cephalomanes atrovirens Presl             <NA>               
-#> 2 54133783 54115097            synonym         species   Trichomanes crassum Copel.                <NA>               
-#> 3 54115097 <NA>                accepted        species   Cephalomanes crassum (Copel.) M. G. Price <NA>               
-#> 4 54133784 54115098            synonym         species   Trichomanes densinervium Copel.           <NA>               
-#> 5 54115098 <NA>                accepted        species   Cephalomanes densinervium (Copel.) Copel. <NA>               
-#> 6 193e2011 <NA>                accepted        <NA>      Hymenophyllum dwctaxonense Nitta          2023-03-21 08:40:04
+#>   taxonID                          acceptedNameUsageID taxonomicStatus taxonRank scientificName                            modified         
+#>   <chr>                            <chr>               <chr>           <chr>     <chr>                                     <chr>            
+#> 1 54115096                         <NA>                accepted        species   Cephalomanes atrovirens Presl             <NA>             
+#> 2 54133783                         54115097            synonym         species   Trichomanes crassum Copel.                <NA>             
+#> 3 54115097                         <NA>                accepted        species   Cephalomanes crassum (Copel.) M. G. Price <NA>             
+#> 4 54133784                         54115098            synonym         species   Trichomanes densinervium Copel.           <NA>             
+#> 5 54115098                         <NA>                accepted        species   Cephalomanes densinervium (Copel.) Copel. <NA>             
+#> 6 193e2011c8ace0ed138af91f41a335cc <NA>                accepted        <NA>      Hymenophyllum dwctaxonense Nitta          2023-04-29 08:53…
 ```
 
 `dct_modify_row()` modifies a row, automatically re-mapping synonyms if
@@ -124,13 +124,13 @@ filmies_small |>
     usage_name = "Cephalomanes crassum (Copel.) M. G. Price"
   )
 #> # A tibble: 5 × 6
-#>   taxonID  acceptedNameUsageID taxonomicStatus taxonRank scientificName                            modified           
-#>   <chr>    <chr>               <chr>           <chr>     <chr>                                     <chr>              
-#> 1 54115096 <NA>                accepted        species   Cephalomanes atrovirens Presl             <NA>               
-#> 2 54133783 54115097            synonym         species   Trichomanes crassum Copel.                <NA>               
-#> 3 54115097 <NA>                accepted        species   Cephalomanes crassum (Copel.) M. G. Price <NA>               
-#> 4 54133784 54115097            synonym         species   Trichomanes densinervium Copel.           2023-03-21 08:40:04
[…]
