+++ README.Rmd
@@ -192,6 +192,7 @@ the right.
 
 Contributions to this package are welcome! Please see the [Contribution Guide](.github/CONTRIBUTING.md) and [Code of Conduct](https://ropensci.org/code-of-conduct/).
 
-## License
+## Licenses
 
-[MIT License](LICENSE.md)
+- All original code is distributed under the [MIT License](LICENSE.md).
+- The `paste3()` function by IRTFM on [Stack Overflow](https://stackoverflow.com/questions/13673894/suppress-nas-in-paste) is used per [CC BY-SA 3.0](https://creativecommons.org/licenses/by-sa/3.0/).
