+++ tests/testthat/test-utils.R
@@ -1,14 +1,3 @@
-test_that("paste3() works", {
-  expect_equal(
-    paste3(1, 2, NA),
-    "1 2"
-  )
-  expect_equal(
-    paste3(1, 2, NA, sep = ", "),
-    "1, 2"
-  )
-})
-
 test_that("drop_first() works", {
   expect_equal(
     drop_first(c(1, 2, 3)),
