+++ R/utils.R
@@ -1,33 +1,3 @@
-#' paste3
-#'
-#' Paste while removing NAs
-#'
-#' Removes NAs from pasted elements, but if ALL elements are NA, the result is
-#' NA.
-#'
-#' Shamelessly copied from
-#' \url{https://stackoverflow.com/questions/13673894/suppress-nas-in-paste}
-#' @param ... Strings to paste
-#' @param sep Character used to separate pasted strings
-#' @noRd
-#' @autoglobal
-paste3 <- function(..., sep = " ") {
-  my_list <- list(...)
-  my_list <- lapply(my_list, function(x) {
-    x[is.na(x)] <- ""
-    x
-  })
-  ret <- gsub(
-    paste0("(^", sep, "|", sep, "$)"), "",
-    gsub(
-      paste0(sep, sep), sep,
-      do.call(paste, c(my_list, list(sep = sep)))
-    )
-  )
-  is.na(ret) <- ret == ""
-  ret
-}
-
 #' Drop the first element of a vector
 #'
 #' @param x Vector
