+++ R/dct_check_mapping_strict.R
@@ -35,15 +35,18 @@ check_mapping_strict_status <- function(tax_dat,
     ignore.case = FALSE
   )
   any_missing <- any(c(
+    valid_tax_status_lacks_syn,
+    valid_tax_status_lacks_acc,
+    valid_tax_status_lacks_var
+  ))
+
+  missing_tax_status <- c("synonym", "accepted", "variant")[
+    c(
       valid_tax_status_lacks_syn,
       valid_tax_status_lacks_acc,
       valid_tax_status_lacks_var
-    ))
-
-  missing_tax_status <- c("synonym", "accepted", "variant")[
-    c(valid_tax_status_lacks_syn,
-      valid_tax_status_lacks_acc,
-      valid_tax_status_lacks_var)]
+    )
+  ]
 
   missing_tax_status <- paste(missing_tax_status, collapse = ", ")
 
