### 2023-05-02 joelnitta/dwctaxon#52 @joelnitta
The [rOpenSci manual](https://devguide.ropensci.org/building.html?#pkgdependencies) states

> Specifying minimum dependencies (e.g. glue (>= 1.3.0) instead of just glue) should be a conscious choice. If you know for a fact that your package will break below a certain dependency version, specify it explicitly. But if you don’t, then no need to specify a minimum dependency. In that case when a user reports a bug which is explicitly related to an older version of a dependency then address it then. An example of bad practice would be for a developer to consider the versions of their current state of dependencies to be the minimal version. That would needlessly force everyone to upgrade (causing issues with other packages) when there is no good reason behind that version choice.

As I do not know of specific versions below which the code breaks I am not going to specify minimum package versions for `Imports`. But if a particular issue pops up where an old version breaks something, I will add it on a case-by-base basis.
