### 2023-05-03 b67d0c4
Allow for self-mapping between taxonID and acceptedNameUsageID for accepted names. Fixes #58
