+++ data-raw/sysdata.R
@@ -1,9 +1,9 @@
 # Set up data for testing check_mapping_to_self()
 bad_dat_dup_taxid <- tibble::tribble(
-  ~taxonID, ~acceptedNameUsageID, ~scientificName,
-  "1", NA, "Species foo",
-  "2", "1", "Species bar",
-  "3", "3", "Species bat"
+  ~taxonID, ~acceptedNameUsageID, ~scientificName, ~taxonomicStatus,
+  "1", NA, "Species foo", "Accepted",
+  "2", "1", "Species bar", "Synonym",
+  "3", "3", "Species bat", "Synonym"
 )
 bad_dat_dup_taxid_cols <- c(
   "acceptedNameUsageID", "parentNameUsageID", "originalNameUsageID"
@@ -35,5 +35,6 @@ usethis::use_data(
   bad_dat_dup_taxid_cols,
   bad_dat_dup_taxid,
   bad_dat_missing_taxid,
-  internal = TRUE
+  internal = TRUE,
+  overwrite = TRUE
 )
