### 2023-05-03 f0accad
Allow for self-mapping between taxonID and acceptedNameUsageID for accepted names. Fixes #58
