### 2023-05-03 894ce64
Add custom checks for dct_options()
