### 2023-05-03 de2f6b2
Add option for extra cols fixes #59
