+++ tests/testthat/test-dct_options.R
@@ -18,6 +18,17 @@ test_that("custom checks on options work", {
     dct_options(valid_tax_status = c("a", "b")),
     "Option value must be a string"
   )
+  expect_error(
+    dct_options(extra_cols = c(1, 2)),
+    "Option value must be a character vector"
+  )
+  expect_error(
+    dct_options(valid_tax_status = NULL),
+    "Option value must be a string"
+  )
+  expect_no_error(
+    dct_options(extra_cols = NULL)
+  )
 })
 
 test_that("dct_options() can be set and unset", {
