+++ R/roxygen.R
@@ -67,6 +67,18 @@ param_valid_tax_status <- glue::glue(
   `"NA"` indicates that missing (NA) values are valid. Case-sensitive.'
 )
 
+# Need to manually specify default value for extra cols (NULL will disappear
+# when glued), so check this assumption:
+if (!is.null(get_dct_opt("extra_cols"))) {
+  stop("Default value for extra_cols should be NULL")
+}
+param_extra_cols <- glue::glue(
+  'Character vector; names of columns that should be allowed beyond
+  those defined by the Darwin Core DWC taxon standard. \\
+  Default NULL. \\
+  Providing column name(s) that are valid DWC taxon column(s) has no effect.'
+)
+
 param_on_fail <- glue::glue(
   'Character vector of length 1, either "error" or "summary". \\
   Describes what to do if the check fails. \\
