+++ _pkgdown.yml
@@ -38,3 +38,4 @@ articles:
   - what-is-dwc
   - editing
   - validation
+  - real-data
