### 2023-05-03 5a7e3c3
Put back real-data vignette Fixes #53, for real
