+++ vignettes/real-data.Rmd
@@ -1,0 +1,197 @@
+---
+title: "Real World Example"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Real World Example}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+
+# Increase width for printing tibbles
+options(width = 140)
+```
+
+This vignette demonstrates using dwctaxon on "real life" data found in the wild. Our goal is to import the data and validate it.
+
+First, load the packages used for this vignette.
+
+```{r setup, message = FALSE}
+library(dwctaxon)
+library(readr)
+library(tibble)
+library(dplyr)
+```
+
+## Import data
+
+We will use the [Database of Vascular Plants of Canada (VASCAN)](http://data.canadensys.net/ipt/resource.do?r=vascan), which is available as a Darwin Core Archive.
+
+The data can be obtained manually by going to the [VASCAN website](http://data.canadensys.net/ipt/resource.do?r=vascan), downloading the Darwin Core Archive, and unzipping it^[If you download the data manually, it may be a different version than the one used here, v37.12].
+
+Alternatively, it can be downloaded and unzipped with R:
+
+```{r download-unzip}
+# Set up folders:
+# - Specify temporary folder for downloading data
+temp_dir <- tempdir()
+# - Set name of zip file
+temp_zip <- paste0(temp_dir, "/dwca-vascan.zip")
+# - Set name of unzipped folder
+temp_unzip <- paste0(temp_dir, "/dwca-vascan")
+
+# Download data
+download.file(
+  url = "https://data.canadensys.net/ipt/archive.do?r=vascan&v=37.12",
+  destfile = temp_zip)
+
+# Unzip
+unzip(temp_zip, exdir = temp_unzip)
+
+# Check the contents of the unzipped data (the Darwin Core Archive)
+list.files(temp_unzip)
+```
+
+Next, load the taxonomic data (`taxon.txt`) into R. It is a tab-separated text file, so we use `readr::read_tsv()` to load it.
+
+```{r load-data}
+vascan <- read_tsv(paste0(temp_unzip, "/taxon.txt"))
+
+# OK to delete the tempo
+
+# Take a peak at the data
+vascan
+```
+
+The dataset includes `r nrow(vascan)` rows (taxa) and `r ncol(vascan)` columns.
+
+# Validation
+
+Let's see if the dataset passes validation with dwctaxon.
+
+It is usually a good idea to just run `dct_validate()` with default settings the first time.
+If it passes, you can move on.
+
+```{r validation, error = TRUE}
+dct_validate(vascan)
+```
+
+Looks like we've got problems...
+
+To dig into these in more detail, let's run `dct_validate()` again, but this time output a summary of errors.
+
+```{r validation-summary}
+validation_res <- dct_validate(vascan, on_fail = "summary")
+
+validation_res
+```
+
+The summary lists one `taxonID` per row. Let's count these to get a higher-level view of what's going on.
+
+```{r summary-analysis}
+validation_res %>%
+  count(check, error)
+```
+
+```{r summary-analysis-hide, show = FALSE, echo = FALSE}
+validation_res_sum <-
+  validation_res %>%
+  count(check, error)
+
+n_error_types <- nrow(validation_res_sum) %>%
+  english::english()
+
+n_bad_cols <- validation_res_sum %>%
+  filter(error == "Invalid column names detected: id") %>%
+  pull(n) %>%
+  english::english()
+
+n_bad_sci_name <- validation_res_sum %>%
[…]
