### 2023-05-03 2af31e9
Add vignette about real world data
fixes #53

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
