+++ pkgdown.yml
@@ -3,9 +3,10 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   editing: editing.html
+  real-data: real-data.html
   validation: validation.html
   what-is-dwc: what-is-dwc.html
-last_built: 2023-05-03T06:19Z
+last_built: 2023-05-03T21:17Z
 urls:
   reference: https://joelnitta.github.io/dwctaxon/reference
   article: https://joelnitta.github.io/dwctaxon/articles
