+++ .pre-commit-config.yaml
@@ -5,22 +5,7 @@ repos:
     rev: v0.3.2.9007
     hooks: 
     -   id: style-files
-        args: [--style_pkg=styler, --style_fun=tidyverse_style]    
-    -   id: roxygenize
-        # roxygen requires loading pkg -> add dependencies from DESCRIPTION
-        additional_dependencies:
-        -    assertr
-        -    assertthat
-        -    digest
-        -    dplyr
-        -    glue
-        -    methods
-        -    purrr
-        -    stringr
-        -    tibble
-        -    tidyr
-        -    utils
-        -    anthonynorth/roxyglobals
+        args: [--style_pkg=styler, --style_fun=tidyverse_style]
     # codemeta must be above use-tidy-description when both are used
     -   id: codemeta-description-updated
     -   id: use-tidy-description
