+++ R/dct_modify_row.R
@@ -70,12 +70,25 @@ dct_modify_row_single <- function(tax_dat,
   assertthat::assert_that(assertthat::is.flag(strict))
   assertthat::assert_that(assertthat::is.flag(quiet))
   # Other input checks ----
+
+  # - scientificName is present if usage_name is specified
+  ifelse(
+    !is.null(usage_name),
+    assertthat::assert_that(
+      isTRUE("scientificName" %in% colnames(tax_dat)),
+      msg = paste(
+        "tax_dat must include column 'scientificName' to look up rows by",
+        "acceptedNameUsage"
+      )
+    ),
+    TRUE
+  )
   # - usage_name must exist in input data
   ifelse(
     !is.null(usage_name),
     assertthat::assert_that(
       usage_name %in% tax_dat$scientificName,
-      msg = "`usage_name` not detected in `tax_dat$scientificName`"
+      msg = "Input acceptedNameUsage not detected in tax_dat$scientificName"
     ),
     TRUE
   )
@@ -84,78 +97,32 @@ dct_modify_row_single <- function(tax_dat,
     !is.null(usage_id),
     assertthat::assert_that(
       usage_id %in% tax_dat$taxonID,
-      msg = "`usage_id` not detected in `tax_dat$taxonID`"
+      msg = "Input acceptedNameUsageID not detected in tax_dat$taxonID"
     ),
     TRUE
   )
 
-  # Convert aliases ----
-  if (!is.null(other_terms)) {
-    if (nrow(other_terms) > 0) {
-      # - taxon_id
-      assertthat::assert_that(
-        sum(!is.null(taxon_id), "taxonID" %in% colnames(other_terms)) != 2,
-        msg = "Must use either taxon_id or taxonID"
+  # - other_terms must be valid DWC terms or an exception specified by
+  # options
+  ifelse(
+    isTRUE(nrow(other_terms) > 0),
+    assertthat::assert_that(
+      all(
+        colnames(other_terms) %in% c(dct_terms$term, dct_options()$extra_cols)
+      ),
+      msg = paste(
+        "All terms to modify must be valid DWC taxon terms (`dct_terms$term`)",
+        "or specified by `dct_options$()extra_cols`"
       )
-      if ("taxonID" %in% colnames(other_terms)) {
-        taxon_id <- other_terms$taxonID[[1]]
-        other_terms$taxonID <- NULL
-      }
-      # - sci_name
-      assertthat::assert_that(
-        sum(
-          !is.null(sci_name), "scientificName" %in% colnames(other_terms)
-        ) != 2,
-        msg = "Must use either sci_name or scientificName"
-      )
-      if ("scientificName" %in% colnames(other_terms)) {
-        sci_name <- other_terms$scientificName[[1]]
-        other_terms$scientificName <- NULL
-      }
-      # - usage_id
-      assertthat::assert_that(
-        sum(
-          !is.null(usage_id),
-          "acceptedNameUsageID" %in% colnames(other_terms)
-        ) != 2,
-        msg = "Must use either usage_id or acceptedNameUsageID"
-      )
-      if ("acceptedNameUsageID" %in% colnames(other_terms)) {
-        usage_id <- other_terms$acceptedNameUsageID[[1]]
-        other_terms$acceptedNameUsageID <- NULL
-      }
-      # - usage_name
-      assertthat::assert_that(
-        sum(
-          !is.null(usage_name),
-          "acceptedNameUsage" %in% colnames(other_terms)
-        ) != 2,
-        msg = "Must use either usage_name or acceptedNameUsage"
-      )
-      if ("acceptedNameUsage" %in% colnames(other_terms)) {
[…]
