+++ inst/examples/dct_add_row.R
@@ -5,7 +5,7 @@ tibble::tibble(
   taxonomicStatus = "accepted"
 ) |>
   dct_add_row(
-    sci_name = "Foogenus barspecies var. bla",
+    scientificName = "Foogenus barspecies var. bla",
     parentNameUsageID = "123",
     nameAccordingTo = "me",
     strict = TRUE
