+++ man/dct_validate.Rd
@@ -48,7 +48,7 @@ dct_validate(
 \item{valid_tax_status}{Character vector of length 1; valid values for \code{taxonomicStatus}. Each value must be separated by a comma. Default \verb{accepted, synonym, variant, NA}. \code{"NA"} indicates that missing (NA) values are valid. Case-sensitive.}
 
 \item{extra_cols}{Character vector; names of columns that should be allowed beyond
-those defined by the Darwin Core (DwC) Taxon standard. Default NULL. Providing column name(s) that are valid DwC taxon column(s) has no effect.}
+those defined by the Darwin Core DWC taxon standard. Default NULL. Providing column name(s) that are valid DWC taxon column(s) has no effect.}
 
 \item{on_success}{Character vector of length 1, either "logical" or "data". Describes what to do if the check passes. Default \code{"data"}.}
 
