+++ R/globals.R
@@ -1,8 +1,6 @@
 # Generated by roxyglobals: do not edit by hand
 
 utils::globalVariables(c(
-  "taxonID", # <dct_add_row>
-  "scientificName", # <dct_add_row>
   "dct_terms", # <check_col_names_p>
   "taxonomicStatus", # <check_syn_map_to_acc>
   "acceptedNameUsageID", # <check_acc_id_has_tax_status>
@@ -13,6 +11,7 @@ utils::globalVariables(c(
   "scientificName", # <check_status_diff_p>
   "taxonomicStatus", # <check_status_diff_p>
   "dct_terms", # <dct_fill_col>
+  "dct_terms", # <dct_modify_row_single>
   "taxonID", # <dct_modify_row_single>
   "scientificName", # <dct_modify_row_single>
   "acceptedNameUsageID", # <dct_modify_row_single>
