+++ tests/testthat/test-dct_modify_row.R
@@ -1,21 +1,22 @@
 # Tests for major arguments ----
 
 library(mockery)
+library(tibble)
 
 # - taxon_id
 # - sci_name
-test_that("sci_name works to identify or modify row", {
-  tax_dat <- tibble::tribble(
+test_that("scientificName works to identify or modify row", {
+  tax_dat <- tribble(
     ~taxonID, ~acceptedNameUsageID, ~taxonomicStatus, ~scientificName,
     "1", NA, "accepted", "foo"
   )
   expect_equal(
     dct_modify_row(
       tax_dat,
-      sci_name = "foo", tax_status = "maybe accepted",
+      scientificName = "foo", taxonomicStatus = "maybe accepted",
       stamp_modified = FALSE
     ),
-    tibble::tribble(
+    tribble(
       ~taxonID, ~acceptedNameUsageID, ~taxonomicStatus, ~scientificName,
       "1", NA, "maybe accepted", "foo"
     )
@@ -23,57 +24,71 @@ test_that("sci_name works to identify or modify row", {
   expect_equal(
     dct_modify_row(
       tax_dat,
-      taxon_id = "1", sci_name = "foobar",
+      taxonID = "1", scientificName = "foobar",
       stamp_modified = FALSE
     ),
-    tibble::tribble(
+    tribble(
       ~taxonID, ~acceptedNameUsageID, ~taxonomicStatus, ~scientificName,
       "1", NA, "accepted", "foobar"
     )
   )
 })
 
-# - tax_status
-# - usage_id
-# - remap_names
-test_that("tax_status, usage_id, and remap_names arguments work", {
-  tax_dat <- tibble::tribble(
-    ~taxonID, ~acceptedNameUsageID, ~taxonomicStatus, ~scientificName,
-    "1", NA, "accepted", "foo",
-    "2", NA, "accepted", "bar",
-    "3", "2", "synonym", "bat"
-  )
-  expect_equal(
+# - ...
+test_that("non-DWC terms fail", {
+  expect_error(
     dct_modify_row(
-      tax_dat,
-      taxon_id = "2", tax_status = "synonym", usage_id = "1",
-      stamp_modified = FALSE
+      data.frame(taxonID = "a"),
+      taxon_id = "a", scientificName = "foobar"
     ),
-    tibble::tribble(
-      ~taxonID, ~acceptedNameUsageID, ~taxonomicStatus, ~scientificName,
-      "1", NA, "accepted", "foo",
-      "2", "1", "synonym", "bar",
-      "3", "1", "synonym", "bat"
-    )
-  )
-  expect_equal(
-    dct_modify_row(
-      tax_dat,
-      taxon_id = "2", tax_status = "synonym", usage_id = "1",
-      stamp_modified = FALSE, remap_names = FALSE
-    ),
-    tibble::tribble(
-      ~taxonID, ~acceptedNameUsageID, ~taxonomicStatus, ~scientificName,
-      "1", NA, "accepted", "foo",
-      "2", "1", "synonym", "bar",
-      "3", "2", "synonym", "bat"
-    )
+    "All terms to modify must be valid DWC taxon terms"
   )
 })
 
+# - taxonomicStatus
+# - acceptedNameUsageID
+# - remap_names
+test_that(
+  "Name remapping works",
+  {
+    tax_dat <- tribble(
+      ~taxonID, ~acceptedNameUsageID, ~taxonomicStatus, ~scientificName,
+      "1", NA, "accepted", "foo",
+      "2", NA, "accepted", "bar",
+      "3", "2", "synonym", "bat"
+    )
+    expect_equal(
+      dct_modify_row(
+        tax_dat,
+        taxonID = "2", taxonomicStatus = "synonym", acceptedNameUsageID = "1",
[…]
