+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   real-data: real-data.html
   validation: validation.html
   what-is-dwc: what-is-dwc.html
-last_built: 2023-05-04T08:13Z
+last_built: 2023-05-07T06:40Z
 urls:
   reference: https://joelnitta.github.io/dwctaxon/reference
   article: https://joelnitta.github.io/dwctaxon/articles
