### 2023-05-07 712d07e
Don't put assertions inside ifelse()
