### 2023-05-08 414e038
Move row isolation to subfunction
