### 2023-05-08 aabe729
Move creation of new row by modification to subfunction
