### 2023-05-08 a2787a8
Move modification of other rows to subfunction
