### 2023-05-08 7789d34
Move formatting of output in dct_modify_row_single to subfunction
