### 2023-05-08 b24b7a6
Make conversion of NA to NULL a subfunction
