### 2023-05-08 80c75f1
Don't put assertions inside ifelse()
