### 2023-05-08 joelnitta/dwctaxon#66 @joelnitta
assertr is causing a dependency but only gets used in one function
As mentioned in [ROpenSci review](https://github.com/ropensci/software-review/issues/574) by @sformel-usgs:

> It appears that the `assertr` package is imported for a single function in `utilties.R`. I don’t know of a simple way to replace this function and reduce the dependency.
