### 2023-05-08 b60ddd8
Delete assert_dat fixes #66
