### 2023-05-08 a433338
Delete assert_dat fixes #66
