### 2023-05-08 joelnitta/dwctaxon#68 @joelnitta
Not clear when new columns are added
As mentioned in [ROpenSci](https://github.com/ropensci/software-review/issues/574) review by @collinschwantes:

> Its not always obvious which functions will generate new columns, and which wont. See `dct_fill_col` vs `dct_add_row`

Assignees:
Labels:
