### 2023-05-08 3e89916
Remove old tests about aliases
