### 2023-05-08 3237aa8
Fix assertion for fill_usage_id in dct_add_row
