### 2023-05-08 495266c
Add stamp_modified arg to dct_fill_col
