### 2023-05-08 414a22c
Add documentation to clarify when new cols are added
