### 2023-05-08 joelnitta/dwctaxon#70 @joelnitta
Feature request: function to drop a row
As mentioned in [ROpenSci review](https://github.com/ropensci/software-review/issues/574) by @collinschwantes:

> Since there is an "add_row" function, it may be worth including a drop_row?

Assignees: @joelnitta
Labels: enhancement
