### 2023-05-08 54ba041
Add function to drop rows fixes #70
