+++ NAMESPACE
@@ -5,6 +5,7 @@ export(dct_check_mapping)
 export(dct_check_sci_name)
 export(dct_check_tax_status)
 export(dct_check_taxon_id)
+export(dct_drop_row)
 export(dct_fill_col)
 export(dct_modify_row)
 export(dct_options)
