### 2023-05-08 c1eff15
Add function to drop rows fixes #70
