### 2023-05-08 #574 @joelnitta
Thanks again to the reviewers for their helpful comments! Here are my responses.

## Responses to @sformel-usgs

> ### Review Comments
>
> I found this package to be thoughtfully written and documented. Code appears to have been carefully constructed and functions well. However, I do have a few high level concerns:
> 1. TL;DR More context and example should be added to the statement of need.

- The Statement of Need has been clarified in the README (https://github.com/joelnitta/dwctaxon/commit/c86673168bfacc6417da57d545f05e8197c8962b).

> 2. I’m not sure why these functions shouldn’t be included as part of the `taxastand` package (or vice versa)? Again, the video referenced above give some sense of why these tasks would be separated, but I think a little more description of how they relate could be useful. I’m not totally convinced that the modification/validation of taxonomic data (`dwctaxon`) should be separated from the standardization of species names from different sources (`taxastand`) as these tasks seem to have a lot of overlap.

- Usage of dwctaxon is not limited to standardization of species names, as described in the revised Statement of Need and demonstrated in the newly added "Real World Data" vignette. Furthermore, different packages / software are available to do name standardization (e.g., [U.Taxonstand](https://doi.org/10.1016/j.pld.2022.09.001) in addition to [taxastand](https://github.com/joelnitta/taxastand)). So in the case of name standardization, I think it makes more since to provide dwctaxon as a separate tool that can be used to prepare the reference database, then the user could standardize names against the reference using the tool of their choice.

> #### Documentation
> 1. Throughout the documentation, but most noticeably in the vignette [‘What is DWC?’ vignette](https://github.com/joelnitta/dwctaxon/blob/main/vignettes/what-is-dwc.Rmd), Darwin Core is abbreviated ‘DWC’. Generally speaking, I see/use ‘DwC’, and that is what is described on the [Wikipedia page](https://en.wikipedia.org/wiki/Darwin_Core) since 2011. Not critical, but I would be curious to hear whether the author has a strong opinion on the appropriate abbreviation. If not, I would prefer DwC.

- "DWC" has been changed to "DwC" throughout (https://github.com/joelnitta/dwctaxon/commit/fd1e14918b0e70f1711584c524e693fd3e1b746a).

> 2. In the [‘What is DWC?’ vignette](https://github.com/joelnitta/dwctaxon/blob/main/vignettes/what-is-dwc.Rmd), line 64 describes `genus`, `family` and `order` as “higher taxonomic levels”. I understand these to be “lower levels”, which seems consistent with the comment in the DwC taxonomic term [higherClassification](https://dwc.tdwg.org/terms/#dwc:higherClassification).

- The wording has been changed to "taxonomic levels above species" (https://github.com/joelnitta/dwctaxon/commit/06964165f1248a2f326e84947be1f6bdafa202dd).

> 3. Line 64 of the [editing vignette](https://github.com/joelnitta/dwctaxon/blob/main/vignettes/editing.Rmd) references ‘hash’ generically. I think it would be worth specifying that it is using MD5.

- It is now specified that MD5 is used (https://github.com/joelnitta/dwctaxon/commit/ef7c3931f3830e5b06a82d5a7535271521f8aa1f).

> 4. `utils.R` credits the `paste3` function to a stackoverflow conversation. I’m not sure of the best way to give credit in this situation, but it seems like it would also be good to include the user [“IRTFM”] and userID [“1855677”], since it is a function that is clearly written by one person.

- `paste3()` is no longer used (https://github.com/joelnitta/dwctaxon/commit/d25180cfd6b23999971c3b8b60cc2aabcd4c41a2).

> #### Logical programming API
>
> 1. Yes, however see suggestions below.
>
> 2. In the function, `dct_add_row`, I’m curious about the decision to only use the first 8 characters from the hash as the taxonID. As you note, this may result in duplicates, why not preserve the entire hash as taxonID?

- Now the entire MD5 is used by default, and there is an option to use fewer characters (https://github.com/joelnitta/dwctaxon/commit/ef7c3931f3830e5b06a82d5a7535271521f8aa1f).

> 3. I disagree with the use/creation of aliases as described in lines 72-89 of the [editing vignette](https://github.com/joelnitta/dwctaxon/blob/main/vignettes/editing.Rmd). This promotes confusion about the name of the DwC standard term. Especially in this case where the term is a function parameter and users can use tab to autocomplete, rather than typing out the names. I think this aspect of the package should be revised, and the vignette can point out tab autocomplete as a way to avoid typing out the rather long names of DwC.

- These aliases are no longer used (https://github.com/joelnitta/dwctaxon/commit/b1196238489e1441c13176c424229dd3c44bf469).

> #### automated tool review
>
> Checking the initial package report generated by our @ropensci-review-bot.
>
> 1. As noted in the `goodpractice` checks, `dct_modify_row` and `dct_add_row` have high-cyclocomplexity. I don’t have any great suggestions for ways to break it up. But, if my suggestion, that aliases for DwC terms should be avoided, is accepted, then some of the checks for misuse of aliases could be removed.

- `dct_modify_row_single()` previously had the highest cyclocomplexity (103). Numerous subfunctions have been split out from `dct_modify_row_single()`, and there are no longer any `if()` statements in the main function, greatly decreasing cyclocomplexity to 14 (https://github.com/joelnitta/dwctaxon/commit/315bd1970cc5dfed9d60b10636735bbc911f6f4e). Unfortunately one of the subfunctions, `create_new_row_by_modification()` still contains a large number of `if()` statements and has high cyclocomplexity (56), but this is a significant improvement from 103. I would like to avoid further breaking up `create_new_row_by_modification()` because it has a well-defined purpose (creating a single row), and it is easier to understand the conditional relationships when they are all at the same level and immediately visible.

> 2. The `library` and `require` calls described by `lintr` in `goodpractices` are acceptable. They all appear to be associated with vignettes, tests, or documentation of raw data creation.
>
> 3. It appears that the `assertr` package is imported for a single function in `utilties.R`. I don’t know of a simple way to replace this function and reduce the dependency.

- Thanks for catching this. That function is no longer needed and has been removed, along with the dependency on `assertr` (https://github.com/joelnitta/dwctaxon/commit/b60ddd8cb57b53218366beb50cf98cf95fa3e7de).

## Responses to @collinschwantes

> ### Review Comments
>
> Overall easy to use and well documented. I really enjoyed testing this package. That being said, the statement of need could more clearly define who the target audience is. Is it biologist who are not necessarily taxonomists, taxonomists, collections managers?

- The target audience is anybody who needs to maintain DwC taxonomic data and uses R. The Statement of Need has been clarified and use-cases added (https://github.com/joelnitta/dwctaxon/commit/c86673168bfacc6417da57d545f05e8197c8962b).

> I tested the package using data from a project I worked on back in 2014. Once I had the taxa data in the proper form, all functions worked as expected. I did notice that certain functions create missing columns or provide errors for non-dct terms. Its not always obvious which functions will generate new columns, and which wont. See `dct_fill_col` vs `dct_add_row`

- Documentation has been added to clarify when new columns are added (https://github.com/joelnitta/dwctaxon/commit/9cb465ed9e65bbb054176f64152fcd1b10880e66). Non-dct terms are never added as new columns.

> * I would prefer that functions use DCW terms instead of abbreviations.
>   I didn't find the terms saved me that much time from a typing perspective
>   and I had to remember the abbreviation plus the actual term. Not a lot of mental effort
>   but non-zero.

- These aliases are no longer used (https://github.com/joelnitta/dwctaxon/commit/b1196238489e1441c13176c424229dd3c44bf469).

> * Since there is an "add_row" function, it may be worth including a drop_row?

- Thanks for the suggestion. This has been added as `dct_drop_row()` (https://github.com/joelnitta/dwctaxon/commit/54ba0419cc74f8cfeb5d0e401d06d4fd54443ca9).

> * `dct_fill_col` does not throw an error if the `fill_from` argument is a non-existent column that is a dct_term.

- Thanks for pointing out this bug, it has been fixed (https://github.com/joelnitta/dwctaxon/commit/3f52a9c1c926a1a42636670f8b3f0c73d5bf2436).

> * One error for failing test when running `devtools::check`. `dct_add_row` throws an error.

- It is not clear to me what error this was, but in its current state `devtools::check()` passes locally and on CI builds.

> * dct_modify_row_single has a lot of if statements that are nicely labelled in the function. Consider whether or not each of those sections of the function could be split into its own, smaller, easier to evaluate function.

- Thanks for the suggestion. Several sections of `dct_modify_row_single()` have been split into subfunctions (https://github.com/joelnitta/dwctaxon/commit/315bd1970cc5dfed9d60b10636735bbc911f6f4e).

> * You could also take advantage of `rlang::empty` to test if something is NA or
>   NULL.

- I have been able to clean up this code by moving the call to `is.null()` as one of the conditionals checked with `assert_that()` instead of running `assert_that()` conditional on the result of `is.null()` (https://github.com/joelnitta/dwctaxon/commit/712d07ea73058ecc2951c3ec6d1c6a7e22a7be62).

> * The dependency section in the packaging guidelines recommends using minimum version numbers
>   for the package dependencies

- The [packaging guidelines](https://devguide.ropensci.org/building.html?#pkgdependencies) only recommend using minimum version numbers if there is a known minimum version that would otherwise cause the package to break. I am not aware of any, so I did not include them.

> * Would be great (though potentially out of scope for the package) to have a
>   vignette that shows how to retrieve taxon data and/or what
>   you can do with taxon data once you have it properly formatted in Darwin core.

- Thanks for the suggestion. This has been added as the "Real World Data" vignette (https://github.com/joelnitta/dwctaxon/commit/2af31e9fa82a2724b24ebad30c3fe86bfbe2e658)
