### 2023-05-09 #574 @sformel-usgs
Thanks @joelnitta , these are all satisfactory responses and revisions.  I appreciate the changes you've made to the Statement of Need and the References in the readme.  The Real World vignette was also a great addition for folks who might be starting out in this realm.

I rechecked the package and only came across one minor issue. On my Windows computer, the real-world data vignette fails to build because the `download.file` call results in a corrupt zip file.  A solution is to add `mode = "wb"`.  But you should double check this, because I'm not 100% that it won't cause problems for non-Windows machines.  This is something new that just started occurring for me (non-text files becoming corrupt). I'm not sure what changed, but apparently, [it's not unusual](https://stackoverflow.com/questions/29814405/why-do-i-need-to-use-mode-wb-with-download-file-for-this-rds-file).
