### 2023-05-09 joelnitta/dwctaxon#73 @joelnitta
Real-world data vignette fails on some Windows computers
Reported by @sformel-usgs in the [rOpenSci review](https://github.com/ropensci/software-review/issues/574#issuecomment-1540136042)

> On my Windows computer, the real-world data vignette fails to build because the download.file call results in a corrupt zip file. A solution is to add mode = "wb".
