### 2023-05-22 #574 @ropensci-review-bot
Approved! Thanks @joelnitta for submitting and @collinschwantes, @sformel-usgs for your reviews! :grin:
