### 2023-05-25 #574 @joelnitta
Thanks again to reviewers @sformel-usgs and @collinschwantes and editor @noamross !

I have completed all of the tasks [above](https://github.com/ropensci/software-review/issues/574#issuecomment-1557409863):

I'm happy this package now has a home at rOpenSci!
